refactor: clean up some util functions
