+++ src/controllers/summarized-county.js
@@ -8,17 +8,17 @@ import {
 import { RESPONSE_TYPES } from '../constants';
 
 import {
-  cleanBodyCreator,
   csvDownloadCreator,
-  csvUploadCreator,
-  getIndexes,
+  extractObjectFieldsCreator,
   getModelAttributes,
+  getModelIndexes,
   getModelNumericAttributes,
   indicatorGeneratorCreator,
   newError,
   offsetYearPassCreator,
   predictionGeneratorCreator,
-  spotCSVUploadCreator,
+  processCSV,
+  processCSVAsync,
   trappingAggregationPipelineCreator,
   upsertOpCreator,
   validateNumberEntry,
@@ -29,11 +29,15 @@ const numericModelAttributes = getModelNumericAttributes(SummarizedCountyModel);
 const spotAttributes = ['state', 'county', 'year', 'spotst0'];
 const downloadFieldsToOmit = ['cleridPerDay', 'spbPerDay'];
 
-// this is a function to clean req.body
-const cleanBody = cleanBodyCreator(modelAttributes);
+/**
+ * @description checks that any provided object contains all the model attributes, and filters out any other values
+ * @param {Object} obj an object to check
+ * @returns {Object|false} the filtered object containing only the model attributes if the provided object contains them, else false
+ */
+const extractModelAttributes = extractObjectFieldsCreator(modelAttributes);
 
 // generic upsert operator for this model
-const upsertOp = upsertOpCreator(getIndexes(SummarizedCountyModel));
+const upsertOp = upsertOpCreator(getModelIndexes(SummarizedCountyModel));
 
 /**
  * @description Fetches one year's data from the summarized county collection.
@@ -80,7 +84,7 @@ export const getByFilter = async (startYear, endYear, state, county) => {
  * @throws RESPONSE_TYPES.BAD_REQUEST if missing input
  */
 export const insertOne = async (body) => {
-  const cleanedBody = cleanBody(body);
+  const cleanedBody = extractModelAttributes(body);
   if (!cleanedBody) throw newError(RESPONSE_TYPES.BAD_REQUEST, 'missing parameter(s) in request body');
 
   const newDoc = new SummarizedCountyModel(cleanedBody);
@@ -171,14 +175,22 @@ const cleanCsv = (row) => {
  * @throws RESPONSE_TYPES.BAD_REQUEST for missing fields
  * @throws other errors depending on what went wrong
  */
-export const uploadCsv = csvUploadCreator(
-  SummarizedCountyModel,
-  cleanCsv,
-  cleanBody,
-  undefined,
-  undefined,
-  upsertOp,
-);
+export const uploadCsv = async (filename) => {
+  const { docs, rowCount } = await processCSV(filename, (row) => {
+    // cast the csv fields to our schema
+    const cleanedData = extractModelAttributes(cleanCsv(row));
+    if (!cleanedData) throw newError(RESPONSE_TYPES.BAD_REQUEST, 'missing fields in csv');
+
+    return cleanedData;
+  });
+
+  // apply another transformation to prepare for upserting
+  const upsertOperations = docs.map(upsertOp);
+  const bulkWriteResult = await SummarizedCountyModel.bulkWrite(upsertOperations);
+
+  console.log(`successfully parsed ${rowCount} rows from csv upload`);
+  return bulkWriteResult;
+};
 
 /**
  * @description cleans row of CSV for spot upload, casts undefined or empty string to null
@@ -202,18 +214,33 @@ const cleanSpotsCsv = (row) => {
 };
 
 /**
- * @description uploads spot data on county level
- * @returns {(filename: String) => Promise} async function receiving filter for data subsetting
+ * @description uploads a csv with spot data to the summarized county collection
+ * @param {String} filename the csv filename on disk
+ * @throws RESPONSE_TYPES.BAD_REQUEST for missing fields
+ * @throws other errors depending on what went wrong
  */
-export const uploadSpotsCsv = spotCSVUploadCreator(
-  SummarizedCountyModel,
-  cleanSpotsCsv,
-  cleanBodyCreator(spotAttributes),
-  undefined,
-  undefined,
-  upsertOp,
-  'county',
-);
+export const uploadSpotsCsv = async (filename) => {
+  const { docs, rowCount } = await processCSVAsync(filename, async (row) => {
+    // cast the csv fields to our schema
+    const cleanedData = extractObjectFieldsCreator(spotAttributes)(cleanSpotsCsv(row));
+    if (!cleanedData) throw newError(RESPONSE_TYPES.BAD_REQUEST, 'missing fields in csv');
+
+    // explicitly look up and set endobrev value for this document
+    const { county, state, year } = cleanedData;
+
+    const matchingDoc = await SummarizedCountyModel.findOne({ state, year, county });
+    const endobrev = matchingDoc?.endobrev || null;
+
+    return { ...cleanedData, endobrev };
+  });
+
+  // apply another transformation to prepare for upserting
+  const upsertOperations = docs.map(upsertOp);
+  const bulkWriteResult = await SummarizedCountyModel.bulkWrite(upsertOperations);
+
+  console.log(`successfully parsed ${rowCount} rows from csv upload`);
+  return bulkWriteResult;
+};
 
 /**
  * @description Summarizes all trapping data at the county level. Will overwrite all entries in this collection.
