+++ src/utils/index.js
@@ -1,13 +1,17 @@
 import {
   cleanCsvCreator,
   csvDownloadCreator,
-  csvUploadCreator,
   deleteFile,
-  getIndexes,
-  spotCSVUploadCreator,
+  processCSV,
+  processCSVAsync,
+} from './csv';
+
+import {
+  getModelAttributes,
+  getModelIndexes,
+  getModelNumericAttributes,
   upsertOpCreator,
-  validateNumberEntry,
-} from './csv-upload';
+} from './mongoose';
 
 import {
   generateResponse,
@@ -16,10 +20,9 @@ import {
 } from './responses';
 
 import {
-  cleanBodyCreator,
-  getModelAttributes,
-  getModelNumericAttributes,
-} from './requests';
+  extractObjectFieldsCreator,
+  validateNumberEntry,
+} from './validators';
 
 import {
   indicatorGeneratorCreator,
@@ -34,32 +37,30 @@ import {
   survey123UnpackCreator,
   survey123WebhookUnpackCreator,
   transformSurvey123GlobalID,
-  unsummarizedDataCsvUploadCreator,
 } from './survey123';
 
 export {
-  cleanBodyCreator,
   cleanCsvCreator,
   csvDownloadCreator,
-  csvUploadCreator,
   csvUploadSurvey123Creator,
   deleteFile,
   deleteInsert,
+  extractObjectFieldsCreator,
   generateErrorResponse,
   generateResponse,
-  getIndexes,
   getModelAttributes,
+  getModelIndexes,
   getModelNumericAttributes,
   indicatorGeneratorCreator,
   newError,
   offsetYearPassCreator,
   predictionGeneratorCreator,
-  spotCSVUploadCreator,
+  processCSV,
+  processCSVAsync,
   survey123UnpackCreator,
   survey123WebhookUnpackCreator,
   transformSurvey123GlobalID,
   trappingAggregationPipelineCreator,
-  unsummarizedDataCsvUploadCreator,
   upsertOpCreator,
   validateNumberEntry,
 };
