+++ src/utils/survey123.js
@@ -203,67 +203,3 @@ export const csvUploadSurvey123Creator = (ModelName, cleanCsv, cleanBody, filter
       });
   });
 };
-
-/**
-* @description higher-order function that creates a csv uploader function for unsummarized data
-* @param {mongoose.Model} ModelName destination Model of upload
-* @param {function} cleanCsv function to cast csv to model schema
-* @param {function} cleanBody function to validate that all model schema parameters are present
-* @param {function} filter optional parameter to conditionally accept documents
-* @param {function} transform optional parameter to modify a document before insertion
-* @param {string} filename csv filename on disk
-* @returns {(filename: String) => Promise}
-* @throws RESPONSE_TYPES.BAD_REQUEST for missing fields
-* @throws other errors depending on what went wrong
-*/
-export const unsummarizedDataCsvUploadCreator = (ModelName, cleanCsv, cleanBody, filter, transform) => async (filename) => {
-  const filepath = path.resolve(__dirname, `../../${filename}`);
-
-  const inserts = [];
-  const deletions = [];
-
-  return new Promise((resolve, reject) => {
-    parseFile(filepath, { headers: true })
-      .on('data', (data) => {
-        // cast the csv fields to our schema
-        const cleanedData = cleanBody(cleanCsv(data));
-        if (!cleanedData) reject(newError(RESPONSE_TYPES.BAD_REQUEST, 'missing fields in csv'));
-
-        // apply filter
-        if (!filter || filter(cleanedData)) {
-          // apply transform
-          const doc = transform ? transform(cleanedData) : cleanedData;
-          inserts.push(doc);
-
-          const { globalID } = doc;
-          deletions.push({ globalID });
-        }
-      })
-      .on('error', (err) => reject(err))
-      .on('end', (rowCount) => {
-        const insertOp = inserts.map((document) => ({
-          insertOne: {
-            document,
-          },
-        }));
-
-        // clear out by globalID
-        const deleteOp = deletions.map(({ globalID }) => ({
-          deleteMany: {
-            filter: { globalID },
-          },
-        }));
-
-        ModelName.bulkWrite(deleteOp, { ordered: false })
-          .then((deleteRes) => {
-            return ModelName.bulkWrite(insertOp, { ordered: false })
-              .then((insertRes) => [deleteRes, insertRes]);
-          })
-          .then((bothRes) => {
-            console.log(`successfully parsed ${rowCount} rows from csv upload`);
-            resolve(bothRes);
-          })
-          .catch((err) => reject(err));
-      });
-  });
-};
